requirements.txt
fastapi
uvicorn
langchain
openai
faiss-cpu
pypdf
python-docx
beautifulsoup4
requests
python-multipart
